adds finished notes for backup section
these notes are rough and the content was broad and complicated. this
section will need additional review before the CKA exam for sure.

Also started the notes for security, but haven't put much into that yet
and didn't think it needed to go in the short commit message. Noting it
here for transparency? accuracy? well, it's here anyway and I wanted to
be thorough. Goodness...